response to CR comment, move HostAddress to plugin api
