+++ server/odc-service/src/main/java/com/oceanbase/odc/service/connection/util/ConnectTypeUtil.java
@@ -29,7 +29,7 @@ import org.apache.commons.lang.StringUtils;
 import com.oceanbase.odc.core.shared.Verify;
 import com.oceanbase.odc.core.shared.constant.ConnectType;
 import com.oceanbase.odc.core.shared.constant.DialectType;
-import com.oceanbase.odc.core.shared.jdbc.HostAddress;
+import com.oceanbase.odc.plugin.connect.api.HostAddress;
 import com.oceanbase.odc.plugin.connect.api.JdbcUrlParser;
 import com.oceanbase.odc.service.plugin.ConnectionPluginUtil;
