+++ server/plugins/connect-plugin-api/src/main/java/com/oceanbase/odc/plugin/connect/api/HostAddress.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.core.shared.jdbc;
+package com.oceanbase.odc.plugin.connect.api;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
