+++ server/plugins/connect-plugin-mysql/src/main/java/com/oceanbase/odc/plugin/connect/mysql/MySQLJdbcUrlParser.java
@@ -26,7 +26,7 @@ import java.util.stream.Collectors;
 import org.mariadb.jdbc.UrlParser;
 import org.mariadb.jdbc.util.Options;
 
-import com.oceanbase.odc.core.shared.jdbc.HostAddress;
+import com.oceanbase.odc.plugin.connect.api.HostAddress;
 import com.oceanbase.odc.plugin.connect.api.JdbcUrlParser;
 
 import lombok.NonNull;
