+++ server/plugins/connect-plugin-ob-mysql/src/main/java/com/oceanbase/odc/plugin/connect/obmysql/OceanBaseJdbcUrlParser.java
@@ -25,7 +25,7 @@ import java.util.stream.Collectors;
 
 import com.oceanbase.jdbc.UrlParser;
 import com.oceanbase.jdbc.util.Options;
-import com.oceanbase.odc.core.shared.jdbc.HostAddress;
+import com.oceanbase.odc.plugin.connect.api.HostAddress;
 import com.oceanbase.odc.plugin.connect.api.JdbcUrlParser;
 
 import lombok.NonNull;
